Create variables.py
This module contains the classes Variables, NumberVariables and
FunctionVariables:

A Variable object is a container with contains a set of variable
identifiers and their values.

NumberVariables is a type of Variable designed to store only
float and int objects.

FunctionVariables is a type of Variable designed to store only
FunctionType objects.