Create evaluator.py
This module provides the evaluate() function to evaluate arithmetic
expressions.

evaluate uses the built-in eval function. For making an optimaze
and safe use of it, eval only has access to _run_function and
to the FunctionVariables instances methods. This prevents eval() from
displaying sensitive data.

An expression can have functions_variables and number_variables from
a FunctionVariables and NumberVariables instance.
For evaluating an expression, all these variables must be replaced with
their respected value or function.

An expression is first analyzed by _replace_number_variables. Each
number_variable in the expression is repleced with it numeric value.

Once the expression does not contain any number_variable, it is
analyzed by _replace_function_variables. Since eval does not have
access to any of the functions stored at function_variables, all
this functions calls are replaced  with a _run_function call.

_run_function runs the given function with the arguments specified.
_run_function is called by eval at evaluate.